fix - data type tests
